+++ src/containers/Home/FavoriteSection.scss
@@ -1,6 +1,6 @@
 @import '../../fonts.scss';
 
-.messages-wrapper {
+.favorite-wrapper {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
@@ -9,29 +9,28 @@
   padding-bottom: 16px;
 }
 
-.messages-wrapper>p {
+.favorite-wrapper>p {
   padding-bottom: 8px;
   white-space: nowrap;
   font-size: 1.5rem;
   cursor: default;
 }
 
-.messages-wrapper>p>i {
+.favorite-wrapper>p>i {
   padding-left: 12px;
   padding-right: 8px;
 }
 
-.messages-wrapper>ul {
+.favorite-wrapper>ul {
   font-size: 1.1rem;
   padding-left: 12px;
-  word-wrap: break-word;
-  word-break: break-word;
+  width: 100%;
 }
 
-.messages-wrapper>ul>li {
+.favorite-wrapper>ul>li {
   display: flex;
   justify-content: flex-start;
-  align-items: center;
+  align-items: flex-start;
   word-wrap: break-word;
   word-break: break-word;
   padding-top: 6px;
