+++ src/containers/Home/GroupDetail.scss
