+++ src/containers/Home/GroupDetailHeader.js
