+++ src/containers/Home/GroupSection.scss
@@ -37,6 +37,7 @@
   padding-left: 12px;
   word-wrap: break-word;
   word-break: break-word;
+  width: 100%;
 }
 
 .group-wrapper>ul>li {
